requirements.txt
# install gpu version
# torch==1.9.1
numpy==1.19.5
pandas==1.1.5
transformers==4.11.3
datasets==1.14.0
pytorch-lightning==1.4.9
rich==10.12.0
wandb==0.12.5
hydra-core==1.1.1
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.1
optuna==2.10.0
plotly==5.3.1
seaborn==0.11.2
matplotlib==3.3.4
scikit-learn==0.24.2
tables==3.6.1
